Renamed http.py to gateways.py
